splitting off logic into methods
moving logic out of commands.py
some other small improvements
